+++ src/panels/lovelace/strategies/get-strategy.ts
@@ -59,7 +59,8 @@ const STRATEGIES: Record<LovelaceStrategyConfigType, Record<string, any>> = {
     light: () => import("../../light/strategies/light-view-strategy"),
     security: () => import("../../security/strategies/security-view-strategy"),
     climate: () => import("../../climate/strategies/climate-view-strategy"),
-    maintenance: () => import("./maintenance/maintenance-view-strategy"),
+    maintenance: () =>
+      import("../../maintenance/strategies/maintenance-view-strategy"),
   },
   section: {
     "common-controls": () =>
